more things moved to lisp.h
