+++ src/lisp.c
@@ -21,84 +21,7 @@
     #error "Unsupported bit width"
 #endif
 
-typedef struct heap
-{
-   cell cells[CELLS];
-   struct heap *next;
-}
-heap;
 
-typedef struct bindFrame
-{
-   struct bindFrame *link;
-   int i, cnt;
-   struct {any sym; any val;} bnd[1];
-}
-bindFrame;
-
-typedef struct inFrame
-{
-   struct inFrame *link;
-   void (*get)(void);
-   FILE *fp;
-   int next;
-}
-inFrame;
-
-typedef struct outFrame
-{
-   struct outFrame *link;
-   void (*put)(int);
-   FILE *fp;
-}
-outFrame;
-
-typedef struct parseFrame
-{
-   int i;
-   uword w;
-   any sym, nm;
-}
-parseFrame;
-
-typedef struct stkEnv
-{
-   cell *stack, *arg;
-   bindFrame *bind;
-   int next;
-   any key, cls, *make, *yoke;
-   inFrame *inFrames;
-   outFrame *outFrames;
-   parseFrame *parser;
-   void (*get)(void);
-   void (*put)(int);
-   bool brk;
-}
-stkEnv;
-
-typedef struct catchFrame
-{
-   struct catchFrame *link;
-   any tag, fin;
-   stkEnv env;
-   jmp_buf rst;
-}
-catchFrame;
-
-typedef struct
-{
-        any sym; any val;
-}
-bindFrameBind;
-
-#define bindFrameSize (sizeof(bindFrame))
-#define bindSize (sizeof(bindFrameBind))
-static inline bindFrame *allocFrame(int l)
-{
-    int s1 = bindFrameSize;
-    int s2 = (l - 1) * bindSize;
-    return (bindFrame*)malloc(s1 + s2);
-};
 
 /*** Macros ***/
 #define Free(p)         ((p)->car=Avail, (p)->cdr=0, (p)->type._t=0,  Avail=(p))
